+++ src/plopp/__init__.py
@@ -16,8 +16,8 @@ backends = BackendManager()
 
 from . import data
 from .core import Node, View, input_node, node, show_graph, widget_node
-from .functions import inspector, plot, scatter3d, slicer, superplot
 from .graphics import Camera, figure1d, figure2d, figure3d, tiled
+from .plotting import inspector, plot, scatter3d, slicer, superplot
 
 
 def patch_scipp():
