+++ src/plopp/core/view.py
@@ -3,6 +3,9 @@
 
 import uuid
 from abc import abstractmethod
+from typing import Any, Dict
+
+import scipp as sc
 
 
 class View:
@@ -23,15 +26,7 @@ class View:
         self.graph_nodes = {}
         for node in nodes:
             node.add_view(self)
-
-    @abstractmethod
-    def notify_view(self, _):
-        """
-        The function that will be called when a parent node is told to notify its
-        children and its views.
-        This has to be overridden by any child class.
-        """
-        return
+        self.artists = {}
 
     @property
     def id(self):
@@ -39,3 +34,34 @@ class View:
         The unique id of the view.
         """
         return self._id
+
+    def notify_view(self, message: Dict[str, Any]):
+        """
+        When a notification is received, request data from the corresponding parent node
+        and update the relevant artist.
+
+        Parameters
+        ----------
+        *message:
+            The notification message containing the node id it originated from.
+        """
+        node_id = message["node_id"]
+        new_values = self.graph_nodes[node_id].request_data()
+        self.update(new_values=new_values, key=node_id)
+
+    @abstractmethod
+    def update(self, new_values: sc.DataArray, key: str, draw: bool):
+        """
+        Update function which is called when a notification is received.
+        This has to be overridden by any child class.
+        """
+        ...
+
+    def render(self):
+        """
+        At the end of figure creation, this function is called to request data from
+        all parent nodes and draw the figure.
+        """
+        for node in self.graph_nodes.values():
+            new_values = node.request_data()
+            self.update(new_values=new_values, key=node.id)
