+++ src/plopp/graphics/__init__.py
@@ -3,10 +3,14 @@
 
 # flake8: noqa E402, F401
 
+from .basefig import BaseFig
 from .camera import Camera
 from .colormapper import ColorMapper
-from .figimage import FigImage
-from .figline import FigLine
+
+# from .figimage import FigImage
+# from .figline import FigLine
 from .figscatter3d import FigScatter3d
 from .figure import figure1d, figure2d, figure3d
+from .lineview import LineView
+from .imageview import ImageView
 from .tiled import tiled
