+++ src/plopp/plotting/inspector.py
