+++ src/plopp/plotting/slicer.py
